+++ Annex/src/utopia/annex/model/schrodinger/CachedFindSchrodinger.scala
@@ -1,8 +1,8 @@
 package utopia.annex.model.schrodinger
 
+import utopia.annex.model.response.RequestNotSent.{RequestSendingFailed, RequestWasDeprecated}
+import utopia.annex.model.response.{RequestResult, Response, ResponseBody}
 import utopia.disciple.model.error.RequestFailedException
-import utopia.annex.model.response.RequestNotSent.RequestFailed
-import utopia.annex.model.response.{RequestNotSent, Response, ResponseBody}
 import utopia.flow.collection.CollectionExtensions._
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -38,35 +38,29 @@ class CachedFindSchrodinger[I](cached: I) extends Schrodinger[Try[I], I]
 	  * @param recordError A function that is called when possible errors are encountered (default = ignore errors)
 	  * @param exc Implicit execution context
 	  */
-	def completeWith(result: Future[Either[RequestNotSent, Response]])(parse: ResponseBody => Try[I])
-					(recordError: Throwable => Unit = _ => ())(implicit exc: ExecutionContext) =
+	def completeWith(result: Future[RequestResult])(parse: ResponseBody => Try[I])
+	                (recordError: Throwable => Unit = _ => ())(implicit exc: ExecutionContext) =
 	{
 		result.onComplete {
 			case Success(result) =>
 				result match {
-					case Right(response) =>
-						response match {
-							case Response.Success(_, body, _) =>
-								val parseResult = parse(body)
-								parseResult.failure.foreach(recordError)
-								complete(parseResult)
-							case Response.Failure(status, message, _) =>
-								val errorMessage = message match {
-									case Some(m) => s"Invitation retrieval failed ($status). Response message: $m"
-									case None => s"Invitation retrieval failed with status $status"
-								}
-								val error = new RequestFailedException(errorMessage)
-								recordError(error)
-								complete(Failure(error))
+					case Response.Success(_, body, _) =>
+						val parseResult = parse(body)
+						parseResult.failure.foreach(recordError)
+						complete(parseResult)
+					case Response.Failure(status, message, _) =>
+						val errorMessage = message match {
+							case Some(m) => s"Invitation retrieval failed ($status). Response message: $m"
+							case None => s"Invitation retrieval failed with status $status"
 						}
-					case Left(notSent) =>
-						notSent match {
-							case RequestFailed(error) =>
-								recordError(error)
-								complete(Failure(error))
-							case _ => complete(Failure(
-								new RequestFailedException("Request was deprecated before it could be sent")))
-						}
+						val error = new RequestFailedException(errorMessage)
+						recordError(error)
+						complete(Failure(error))
+					case RequestWasDeprecated =>
+						complete(Failure(new RequestFailedException("Request was deprecated before it could be sent")))
+					case RequestSendingFailed(error) =>
+						recordError(error)
+						complete(Failure(error))
 				}
 			case Failure(error) =>
 				recordError(error)